Move all AWS files to dedicated AWS directory
